+++ service/controller/resource/tccpn/template/template_main_etcd_volume.go
@@ -16,6 +16,12 @@ const TemplateMainEtcdVolume = `
       - Key: Name
         Value: {{ .Name }}
       VolumeType: gp3
+      {{- if and (ge .Iops 3000) (le .Iops 16000) }}
+      Iops: {{ .Iops }}
+      {{- end }}
+      {{- if and (ge .Throughput 125) (le .Throughput 1000) }}
+      Throughput: {{ .Throughput}}
+      {{- end }}
 {{- end -}}
 {{- end -}}
 `