+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Allow resource limits/requests to be passed as values.
 - Switch `gp2` to `gp3` volumes.
+- Allow etcd volume IOPS and Throughput to be set.
 
 ## [10.18.0] - 2022-03-04